refactoring sky subtraction
