+++ src/motes/common.py
@@ -393,132 +393,6 @@ def interpolate_extraction_lims(extraction_limits, dispersion_axis_length):
     return final_extraction_limits
 
 
-# Takes a data column, spatial axis and seeing of the observation and fits a Moffat function to the
-# column using a Levenberg-Marquardt least squares method. Returns the best fit parameters of the
-# Moffat function
-def linear_least_squares(r, col):
-    """
-    Fit a linear profile to a data column using a Levenberg-Marquardt least squares method.
-
-    Args:
-        r (numpy.ndarray)   : Spatial axis of the data column being fitted.
-        col (numpy.ndarray) : Data column being fitted.
-
-    Returns:
-        [res_lsq.x[0], res_lsq.x[1]] (list) : list containing the best fit parameters of the linear
-                                              profile.
-    """
-    # Set up initial conditions for the least squares fit.
-    x0 = [0.0, np.median(col)]
-
-    # Run the least squares fit.
-    res_lsq = least_squares(linear_resid, x0, args=(r, col), method="trf")
-
-    return [res_lsq.x[0], res_lsq.x[1]]
-
-
-def linear_resid(x, data_range, data):
-    """
-    Calculate residuals of fitted linear profile and the data for the Levenberg-Marquardt least squares method.
-
-    Args:
-        x (list)                  : A list containing the best fit parameters of the linear
-                                    profile.
-        data_range (numpy.ndarray) : the spatial axis of the data column.
-        data (numpy.ndarray)      : the data column.
-
-    Returns:
-        residual (numpy.ndarray)  : The residuals of the fitted line and the data.
-    """
-    residual = (x[0] * data_range) + x[1] - data
-    return residual
-
-
-def poly2_least_squares(r, col):
-    """
-    Fits a second-order polynomial to a data column using a Levenberg-Marquardt least squares
-    method.
-
-    Args:
-        r (numpy.ndarray)   : Spatial axis of the data column being fitted.
-        col (numpy.ndarray) : Data column being fitted.
-
-    Returns:
-       params (list)        : the parameters of the fitted polynomial.
-    """
-
-    # Set up initial conditions for the least squares fit.
-    x0 = [0.0, 0.0, np.median(col)]
-
-    # Run the least squares fit.
-    res_lsq = least_squares(poly2_resid, x0, args=(r, col), method="trf")
-
-    params = [res_lsq.x[0], res_lsq.x[1], res_lsq.x[2]]
-    return params
-
-
-def poly2_resid(x, data_range, data):
-    """
-    Calculates residuals of a fitted second-order polynomial and the data for the
-    Levenberg-Marquardt least squares method.
-
-    Args:
-        x (list)                  : A list containing the best fit parameters of the polynomial.
-        data_range (numpy.ndarray) : the spatial axis of the data column.
-        data (numpy.ndarray)      : the data column.
-
-    Returns:
-        residual (numpy.ndarray) : The residuals of the fitted polynominal and the data.
-    """
-    residual = (x[0] * data_range * data_range) + (x[1] * data_range) + x[2] - data
-
-    return residual
-
-
-def poly3_least_squares(r, col):
-    """
-    Fits a third-order polynomial to a data column using a Levenberg-Marquardt least squares
-    method.
-
-    Args:
-        r (numpy.ndarray)   : Spatial axis of the data column being fitted.
-        col (numpy.ndarray) : Data column being fitted.
-
-    Returns:
-        params (list) : the best-fit parameters of the fitted polynomial.
-    """
-    # Set up initial conditions for the least squares fit.
-    x0 = [0.0, 0.0, 0.0, np.median(col)]
-
-    # Run the least squares fit.
-    res_lsq = least_squares(poly3_resid, x0, args=(r, col), method="trf")
-
-    return [res_lsq.x[0], res_lsq.x[1], res_lsq.x[2], res_lsq.x[3]]
-
-
-def poly3_resid(x, data_range, data):
-    """
-    Calculates residuals of a fitted third-order polynomial and the data for the
-    Levenberg-Marquardt least squares method.
-
-    Args:
-        x (list)                  : A list containing the best fit parameters of the polynomial.
-        data_range (numpy.ndarray) : the spatial axis of the data column.
-        data (numpy.ndarray)      : the data column.
-
-    Returns:
-        residual (numpy.ndarray) : The residuals of the fitted polynominal and the data.
-    """
-    residual = (
-        (x[0] * data_range * data_range * data_range)
-        + (x[1] * data_range * data_range)
-        + (x[2] * data_range)
-        + x[3]
-        - data
-    )
-    return residual
-
-
 def make_wav_axis(start, increment, length):
     """
     Returns a wavelength axis array using a start wavelength, the wavelength increment and the
@@ -911,7 +785,7 @@ def set_extraction_limits(moffat_parameters, width_multiplier=3.0):
     lower_extraction_limit = moffat_parameters[1] - (width_multiplier * fwhm)
     upper_extraction_limit = moffat_parameters[1] + (width_multiplier * fwhm)
 
-    return lower_extraction_limit, upper_extraction_limit, moffat_parameters[1]
+    return [lower_extraction_limit, upper_extraction_limit, moffat_parameters[1]]
 
 
 def show_img(data_2d, axes_dict, header_parameters, draw_lines, title):
@@ -1028,212 +902,3 @@ def show_img(data_2d, axes_dict, header_parameters, draw_lines, title):
         plt.show()
 
     return None
-
-
-def subtract_sky(background_spatial_lo_limit, background_spatial_hi_limit, frame_dict, axes_dict, parameters, header_parameters):
-    """
-    Subtracts the sky background from the 2D image by defining bg regions using limits input to the
-    function and then fitting a profile to the background column by column while masking cosmic
-    rays. The background level or profile for each column is subtracted from the full column to
-    produce a background subtracted 2D image.
-
-    Args:
-        background_spatial_lo_limit (numpy.ndarray)  : Lower limits of the background regions.
-        background_spatial_hi_limit (numpy.ndarray) : Upper limits of the background regions.
-        frame_dict (dict)              : A dictionary containing the 2D image.
-        axes_dict (dict)             : A dictionary containing the axis information.
-        parameters (dict)               : A dictionary containing MOTES parameters.
-        header_parameters (dict)              : A dictionary containing header information.
-
-    Returns:
-        frame_dict (dict) : Dictionary containing the background subtracted 2D image.
-    """
-
-    frame_dict["data"] = frame_dict["data"].T
-    frame_dict["errs"] = frame_dict["errs"].T
-
-    sky_model = []
-    number_of_columns = len(background_spatial_lo_limit)
-
-    # Makes sure the limits are within the image.
-    for ii in range(number_of_columns):
-        if background_spatial_lo_limit[ii] < 0:
-            background_spatial_lo_limit[ii] = 0
-        if background_spatial_hi_limit[ii] > (axes_dict["spatial_axis"] + axes_dict["data_spatial_floor"])[-1]:
-            background_spatial_hi_limit[ii] = (axes_dict["spatial_axis"] + axes_dict["data_spatial_floor"])[-1]
-
-        data_column = frame_dict["data"][ii]
-        column_axis = np.array(range(len(data_column)))
-        sky_pixels = data_column[
-            np.where(np.logical_or(column_axis < background_spatial_lo_limit[ii], column_axis > background_spatial_hi_limit[ii]))
-        ]
-
-        # Kill MOTES if there isn't enough background sky to perform sky subtraction. Should
-        # probably be made more nuanced later.
-        if len(sky_pixels) == 0:
-            sys.stdout.write(" >>> No pixels contained inside sky region.\n")
-            sys.stdout.write(
-                "     -BG_FWHM_MULTIPLIER in motesparams.txt is probably too large.\n"
-            )
-            sys.stdout.write("     Please reduce -BG_FWHM_MULTIPLIER and try again.\n")
-            sys.stdout.write(
-                "     -BG_FWHM_MULTIPLIER < "
-                + str(
-                    round(
-                        np.min(
-                            [
-                                np.shape(frame_dict["data"])[0],
-                                np.shape(frame_dict["data"])[1],
-                            ]
-                        )
-                        / (2 * header_parameters["seeing"]),
-                        1,
-                    )
-                )
-                + " recommended in this case.\n"
-            )
-            sys.stdout.write(
-                "     Enlarging the 2D spectrum region in reg.txt is also a viable solution.\n"
-            )
-            sys.stdout.write("     Terminating MOTES.\n")
-            sys.exit()
-
-        sky_axis = column_axis[
-            np.where(np.logical_or(column_axis < background_spatial_lo_limit[ii], column_axis > background_spatial_hi_limit[ii]))
-        ]
-        if len(set(sky_pixels)) == 1:
-            continue
-
-        sky_pixels_median = np.nanmedian(sky_pixels)
-        sky_pixels_sigma = np.nanstd(sky_pixels)
-        good_sky_pixels_location = np.where(
-            np.logical_and(
-                sky_pixels > sky_pixels_median - (10 * sky_pixels_sigma),
-                sky_pixels < sky_pixels_median + (10 * sky_pixels_sigma),
-            )
-        )
-        good_sky_pixels = sky_pixels[good_sky_pixels_location]
-        good_sky_axis = sky_axis[good_sky_pixels_location]
-
-        if parameters["-SKYSUB_MODE"] == "MEDIAN":
-            bootstrapped_sky_pixels = np.random.choice(good_sky_pixels, (len(good_sky_pixels), 100), replace=True)
-            median_sky_sample = np.nanmedian(bootstrapped_sky_pixels, axis=0)
-            column_sky_model = np.nanmean(median_sky_sample)
-            sky_model.append(column_sky_model)
-            sky_model_err = np.std(median_sky_sample) / (99**0.5)
-
-        if parameters["-SKYSUB_MODE"] == "LINEAR":
-            bootstrapped_sky_pixels = np.random.choice(good_sky_pixels, (len(good_sky_pixels), 100), replace=True)
-            linear_terms = []
-            intercepts = []
-            for jj in bootstrapped_sky_pixels.T:
-                sky_model_parameters = linear_least_squares(good_sky_axis, jj)
-                linear_terms.append(sky_model_parameters[0])
-                intercepts.append(sky_model_parameters[1])
-            intercepts = np.array(intercepts)
-            linear_terms = np.array(linear_terms)
-            sky_linear_term = np.mean(linear_terms)
-            sky_linear_term_err = np.std(linear_terms) / (99**0.5)
-            sky_intercept = np.mean(intercepts)
-            sky_intercept_err = np.std(intercepts) / (99**0.5)
-            column_sky_model = (sky_linear_term * column_axis) + sky_intercept
-            sky_model.append(column_sky_model)
-            sky_model_err = (
-                (sky_linear_term_err * column_axis * sky_linear_term_err * column_axis)
-                + (sky_intercept_err * sky_intercept_err)
-            ) ** 0.5
-
-        if parameters["-SKYSUB_MODE"] == "POLY2":
-            bootstrapped_sky_pixels = np.random.choice(good_sky_pixels, (len(good_sky_pixels), 100), replace=True)
-            linear_terms = []
-            intercepts = []
-            quadratic_terms = []
-            for jj in bootstrapped_sky_pixels.T:
-                sky_model_parameters = poly2_least_squares(good_sky_axis, jj)
-                quadratic_terms.append(sky_model_parameters[0])
-                linear_terms.append(sky_model_parameters[1])
-                intercepts.append(sky_model_parameters[2])
-            quadratic_terms = np.array(quadratic_terms)
-            intercepts = np.array(intercepts)
-            linear_terms = np.array(linear_terms)
-            sky_quadratic_term = np.mean(quadratic_terms)
-            sky_quadratic_term_err = np.std(quadratic_terms) / (99**0.5)
-            sky_linear_term = np.mean(linear_terms)
-            sky_linear_term_err = np.std(linear_terms) / (99**0.5)
-            sky_intercept = np.mean(intercepts)
-            sky_intercept_err = np.std(intercepts) / (99**0.5)
-            column_sky_model = (sky_quadratic_term * column_axis * column_axis) + (sky_linear_term * column_axis) + sky_intercept
-            sky_model.append(column_sky_model)
-            sky_model_err = (
-                (sky_quadratic_term_err * sky_quadratic_term_err * column_axis * column_axis * column_axis * column_axis)
-                + (sky_linear_term_err * column_axis * sky_linear_term_err * column_axis)
-                + (sky_intercept_err * sky_intercept_err)
-            ) ** 0.5
-
-        if parameters["-SKYSUB_MODE"] == "POLY3":
-            bootstrapped_sky_pixels = np.random.choice(good_sky_pixels, (len(good_sky_pixels), 100), replace=True)
-            linear_terms = []
-            intercepts = []
-            quadratic_terms = []
-            cubic_terms = []
-            for jj in bootstrapped_sky_pixels.T:
-                sky_model_parameters = poly3_least_squares(good_sky_axis, jj)
-                cubic_terms.append(sky_model_parameters[0])
-                quadratic_terms.append(sky_model_parameters[1])
-                linear_terms.append(sky_model_parameters[2])
-                intercepts.append(sky_model_parameters[3])
-            cubic_terms = np.array(cubic_terms)
-            quadratic_terms = np.array(quadratic_terms)
-            intercepts = np.array(intercepts)
-            linear_terms = np.array(linear_terms)
-            sky_cubic_term = np.mean(cubic_terms)
-            sky_cubic_term_err = np.std(cubic_terms) / (99**0.5)
-            sky_quadratic_term = np.mean(quadratic_terms)
-            sky_quadratic_term_err = np.std(quadratic_terms) / (99**0.5)
-            sky_linear_term = np.mean(linear_terms)
-            sky_linear_term_err = np.std(linear_terms) / (99**0.5)
-            sky_intercept = np.mean(intercepts)
-            sky_intercept_err = np.std(intercepts) / (99**0.5)
-            column_sky_model = (
-                (sky_cubic_term * column_axis * column_axis * column_axis)
-                + (sky_quadratic_term * column_axis * column_axis)
-                + (sky_linear_term * column_axis)
-                + sky_intercept
-            )
-            sky_model.append(column_sky_model)
-            sky_model_err = (
-                (
-                    sky_cubic_term_err
-                    * sky_cubic_term_err
-                    * column_axis
-                    * column_axis
-                    * column_axis
-                    * column_axis
-                    * column_axis
-                    * column_axis
-                )
-                + (sky_quadratic_term_err * sky_quadratic_term_err * column_axis * column_axis * column_axis * column_axis)
-                + (sky_linear_term_err * column_axis * sky_linear_term_err * column_axis)
-                + (sky_intercept_err * sky_intercept_err)
-            ) ** 0.5
-
-        frame_dict["data"][ii] -= column_sky_model
-        frame_dict["errs"][ii] = ((frame_dict["errs"][ii] ** 2) + (sky_model_err**2)) ** 0.5
-
-        sys.stdout.write(
-            "     " + str(ii + 1) + "/" + str(number_of_columns) + " columns completed.\r"
-        )
-
-    sky_model = np.array(sky_model)
-    if parameters["-SKYSUB_MODE"] == "MEDIAN":
-        frame_dict["sky_model"] = np.tile(sky_model, (np.shape(frame_dict["data"])[1], 1))
-    else:
-        frame_dict["sky_model"] = sky_model.T
-
-    frame_dict["data"] = frame_dict["data"].T
-    frame_dict["errs"] = frame_dict["errs"].T
-
-    chip_gaps = np.where(np.nanmedian(frame_dict["data"], axis=0) == 0)
-    frame_dict["data"][:, chip_gaps[0]] += 1
-
-    return frame_dict
